requirements.txt
# ── Core Dash ──────────────────────────────────────────────
dash==2.17.1
dash-bootstrap-components==1.6.0
flask==3.0.3
plotly==5.22.0

# ── Data ───────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.3

# ── Machine Learning ───────────────────────────────────────
scikit-learn==1.5.2

# ── MongoDB ────────────────────────────────────────────────
pymongo==4.8.0

# ── PDF & Scraping ─────────────────────────────────────────
pdfplumber==0.11.4
requests==2.32.3

# ── Export rapports ────────────────────────────────────────
reportlab==4.2.2
xlsxwriter==3.2.0

# ── Utilitaires ────────────────────────────────────────────
python-dotenv==1.0.1
gunicorn==22.0.0